+++ uninstall.sh
@@ -212,14 +212,18 @@ safe_uninstall_mosdns() {
 }
 
 safe_uninstall_daed() {
+    PKG_MGR="$1"
     log "开始安全卸载 daed"
 
+    stop_disable_service luci_daed
     stop_disable_service daed
-    remove_paths /usr/bin/daed /usr/share/daed /etc/init.d/daed
+    remove_pkg_if_installed "$PKG_MGR" luci-i18n-daed-zh-cn
+    remove_pkg_if_installed "$PKG_MGR" luci-app-daed
+    remove_paths /usr/bin/daed /usr/share/daed /etc/init.d/daed /var/log/daed
 
     if [ "$DELETE_CONFIG" -eq 1 ]; then
-        log "删除 daed 配置目录"
-        remove_paths /etc/daed
+        log "删除 daed 配置目录和 LuCI 配置"
+        remove_paths /etc/daed /etc/config/daed
     else
         warn "默认保留 /etc/daed 配置目录"
     fi
@@ -308,7 +312,7 @@ main() {
             safe_uninstall_mosdns "$PKG_MGR"
             ;;
         daed)
-            safe_uninstall_daed
+            safe_uninstall_daed "$PKG_MGR"
             ;;
         openclash)
             safe_uninstall_openclash "$PKG_MGR"